data_simulator/requirements.txt
pandas
numpy
faker
google-cloud-bigquery
pyarrow

